requirements.txt
pyowm~=3.3.0
discord.py>=2.3.2
aiosqlite~=0.19.0